requirements.txt
crewai>=0.102.0
python-dotenv>=1.0.1
requests>=2.32.3
